requirements.txt
altair==5.5.0
joblib==1.3.2
matplotlib==3.8.4
pandas==2.2.2
Pillow==10.2.0
plotly==5.9.0
scikit-learn>1.0.2
streamlit==1.29.0
fastapi
uvicorn[standard]
python-dotenv
